pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.75.0",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "docx2txt>=0.9",
    "fastapi>=0.127.0",
    "fpdf>=1.7.2",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "litellm>=1.80.11",
    "pandas>=2.3.3",
    "pdfminer-six>=20251107",
    "pdfplumber>=0.11.8",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pyspellchecker>=0.8.4",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.21",
    "reportlab>=4.4.7",
    "resend>=2.19.0",
    "slowapi>=0.1.9",
    "spacy>=3.8.11",
    "supabase>=2.27.0",
    "uvicorn>=0.40.0",
]
